+++ spinta/manifests/tabular/helpers.py
@@ -2603,6 +2603,8 @@ def datasets_to_tabular(
                         access=access,
                         order_by=order_by,
                     )
+                    for resource in get_resources_without_models(dataset):
+                        yield from _resource_to_tabular(resource)
             elif dataset is not None and model.external.dataset is None:
                 dataset = None
                 resource = None
@@ -2962,3 +2964,7 @@ def _get_state_obj(reader: TabularReader) -> Optional[TabularReader]:
         return None
     else:
         return reader
+
+
+def get_resources_without_models(dataset: Dataset) -> list[Resource]:
+    return [resource for resource in dataset.resources.values() if not resource.models]