+++ spinta/types/config.py
@@ -86,6 +86,7 @@ def load(context: Context, config: Config) -> Config:
     config.root = rc.get("root", default=None)
     config.max_api_file_size = rc.get("max_file_size", default=100)
     config.max_error_count_on_insert = rc.get("max_error_count_on_insert", default=100)
+    config.load_backends = rc.get("load_backends", default=True)
     if config.root is not None:
         config.root = config.root.strip().strip("/")
 
@@ -97,6 +98,8 @@ def load(context: Context, config: Config) -> Config:
             "Configuration option `mode` must be added to a manifest, now it is added to the config root."
         )
     config.upgrade_mode = rc.get("upgrade_mode", default=False)
+
+    config.cache_control = rc.get("cache_control_header", default="")
 
     return config