+++ CHANGES.rst
@@ -1,8 +1,26 @@
 Changes
 #######
 
-0.2dev6 (unreleased)
+0.2dev7 (unreleased)
+===================
+
+Bug fixes:
+
+- Removed `_base` column from HTML response when viewing SOAP data with URL parameters (`#1338`_)
+- Added required parameters validation, when building SOAP query, and raising exception `MissingRequiredProperty` if parameter is missing (`#1338`_)
+
+  .. _#1338: https://github.com/atviriduomenys/spinta/issues/1338
+
+
+0.2dev6 (2025-10-09)
 ====================
+
+New Features:
+
+- Added config based OpenAPI generator that creates REST API documentation from manifest. (`#1463`_)
+
+
+  .. _#1463: https://github.com/atviriduomenys/spinta/issues/1463
 
 Improvements:
 
@@ -14,6 +32,7 @@ Improvements:
 - Calling getall with soap backend will always return list of object, even if SOAP response has one element (`#1486`_)
 - Introduce `base64()` prepare function that decodes base64 string (`#1486`_)
 - `base64()` prepare function that decodes base64 binary now does this outside dask dataframes. Also, raises `InvalidBase64String` for invalid base64 (`#1486`_)
+- Added `Cache-Control`, `ETag` and `Last-Modified` headers to most commonly used `GET` requests (`#1506`_).
 
   .. _#605: https://github.com/atviriduomenys/spinta/issues/605
   .. _#675: https://github.com/atviriduomenys/spinta/issues/675
@@ -21,14 +40,22 @@ Improvements:
   .. _#1462: https://github.com/atviriduomenys/spinta/issues/1462
   .. _#1486: https://github.com/atviriduomenys/spinta/issues/1486
 
-Other:
-- Removed dependency `mypy`
-
 Bug fixes:
 
+- Fixed a bug where `spinta` was trying to connect to a wsdl source during `spinta check` (`#1424`_).
+- Fixed `spinta copy` ignores resources without any models (`#1512`_)
 - Fixed bugs in `spinta copy` and `spinta check` commands where properties starting with underscore symbol were omitted or caused errors. (`#963`_)
 
+  .. _#1512: https://github.com/atviriduomenys/spinta/issues/1512
+  .. _#1424: https://github.com/atviriduomenys/spinta/issues/1424
   .. _#963: https://github.com/atviriduomenys/spinta/issues/963
+
+
+Other:
+
+- Removed dependency `mypy`
+
+
 
 0.2dev5 (2025-09-03)
 ====================
@@ -62,6 +89,7 @@ New Features:
 - Sprint Review fixes Part 1: Create data service following the Agent name; Remove distribution creation; Try to retrieve Data service before creating one. (`#1415`_)
 - Sprint Review fixes Part 2: Generate dataset name from title or from the last part of dataset column value; Hide `visibility=private` rows; Add the full Dataset name in the DSA. (`#1415`_)
 - Sprint Review fixes Part 2.1: Adjust docstrings. (`#1415`_)
+
   .. _#1274: https://github.com/atviriduomenys/spinta/issues/1274
   .. _#1275: https://github.com/atviriduomenys/spinta/issues/1275
   .. _#1378: https://github.com/atviriduomenys/spinta/issues/1378